requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
itypes==1.2.0
mysql-connector-python==8.0.32
protobuf==3.20.3
python-telegram-bot==20.1
pytz==2022.4
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
